requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
exceptiongroup==1.1.2
fastapi==0.101.0
fpdf==1.7.2
googlemaps==4.10.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
motor==3.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.5.0
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
reportlab==4.0.4
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
